Update README.md
In step 11 of the README, the commmands to store the CHUCKBOT and
MOVIEBOT ARN were failing. This was only occuring if the
user's aws-cli config was setup to use something other than
JSON for the default output. To fix the issue I changed the commands
to explicitly use json for the output type.